Move custom spec hook test to registry suite
